Add pytest-cov and extract shared AWS IoT test mocks
- Move FakeMqttClient, FakeThings and the manager-builder helpers from test_microwave.py to tests/awsiot/mocks.py
- Move inline microwave STATE/THING data and the capability JSONs to tests/data/awsiot/
- Add pytest-cov to requirements-dev.txt
