+++ tests/data/awsiot/capability_mwo.json
