+++ tests/data/awsiot/capability_mwo_no_hood.json
