+++ tests/data/awsiot/microwave_thing.json
@@ -1,0 +1,10 @@
+{
+  "thingName": "WPR1A00000001",
+  "thingTypeName": "KMMC5019JBS",
+  "attributes": {
+    "Name": "4d79204d6963726f77617665",
+    "Category": "Cooking",
+    "Serial": "D1",
+    "CapabilityPartNumber": "W11788386"
+  }
+}
